Refactorización de la capa de servicios para normalizar la tabla de la base de datos. Se ha introducido el objeto de transferencia de datos StudentCourseDTO

+++ server/src/main/java/com/github/angel/raa/modules/service/intefaces/NotaService.java
@@ -9,6 +9,6 @@ public interface NotaService {
     List<NotaDTO> getNotaByStudent();
     Response save(NotaDTO dto);
     Response delete(Long id);
-    Response update(Long notaId, int qualification);
+    Response update(Long notaId, Long qualification);
 
 }
